+++ src/lib.rs
@@ -1,7 +1,7 @@
-//! # [Anor](https://github.com/anor-rs)
-//! 
-//! **The open-source data storage platform**
+//! The Storage API Service for [Anor](https://github.com/anor-rs)
 //! 
 //! ## Project Stage
 //! 
 //! **Research:** This project is at the design stage, with some sketches of work but nothing usable yet.
+
+pub mod storage_api;