+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+* added socket server
+
 ## 0.1.0 (2023-10-19)
 
 * project started